+++ ici/library/platform.c
@@ -1566,19 +1566,19 @@ void	writeErrMemo(char *text)
 	writeMemoNote(text, system_error_msg());
 }
 
-char	*iToa(long long arg)
+char	*iToa(int arg)
 {
 	static ION_THREAD_LOCAL char itoa_str[33];
 
-	isprintf(itoa_str, sizeof itoa_str, "%lld", arg);
+	isprintf(itoa_str, sizeof itoa_str, "%d", arg);
 	return itoa_str;
 }
 
-char	*uToa(unsigned long long arg)
+char	*uToa(unsigned int arg)
 {
 	static ION_THREAD_LOCAL char utoa_str[33];
 
-	isprintf(utoa_str, sizeof utoa_str, "%llu", arg);
+	isprintf(utoa_str, sizeof utoa_str, "%u", arg);
 	return utoa_str;
 }
